+++ src/main/java/com/project/dao/CategoryExpensesDao.java
@@ -8,7 +8,7 @@ import com.project.domain.CategoryRevenueEntity;
 
 public interface CategoryExpensesDao extends GenericDao<CategoryExpensesEntity, String>{
 	public List<CategoryExpensesEntity> getAllCategoryRevenuesForUser(String username);
-	public Integer getCategoryIdByName(String name);
+	public Long getCategoryIdByName(String name);
 	public CategoryExpensesEntity getCategoryByName(String name);
 	public boolean daleteRevenueCategoryByName(String name);
 }
