+++ src/main/java/com/project/services/CategoryService.java
@@ -10,13 +10,13 @@ public interface CategoryService {
 	CategoryRevenueEntity getCategoryRevenueEntity(String categoryName);
 	CategoryRevenueEntity updateCategoryRevenueEntity(CategoryRevenueEntity categoryRevenueEntity);
 	public List<CategoryRevenueEntity> getAllCategoryRevenuesForUser(String username);
-	public Integer getCategoryRevenueIdByName(String name);
+	public Long getCategoryRevenueIdByName(String name);
 	public boolean daleteRevenueCategoryByName(String name);
 	
 	CategoryExpensesEntity careateNewExpensesCategory(CategoryExpensesEntity categoryRevenueEntity);
 	CategoryExpensesEntity getCategoryExpensesEntity(String categoryName);
 	CategoryExpensesEntity updateCategoryExpensesEntity(CategoryExpensesEntity categoryRevenueEntity);
 	public List<CategoryExpensesEntity> getAllCategoryExpensesForUser(String username);
-	public Integer getCategoryExpensesIdByName(String name);
+	public Long getCategoryExpensesIdByName(String name);
 	public boolean daleteExpensesCategoryByName(String name);
 }
